Turn components' fields list into actual list
The previous fields "list" was actually an Object type, but
the order of keys is important. This also makes the
"generic" parsing function easier to read.

Add types for all of the possible fields configurations.

Add Component type as return type for convert_* functions.

Fix bug in BufferSave's BufferNum util function.