Type Component and its output values
Add types for a webvs component and most field value types.
Apply these types to the appropriate util functions.

Split util.ts:callFunction into callFunction (with 3 or 4
parameters) and callPostProcFunction (with only 2
parameters). The latter is used in a few cases to convert
the output of a previous callFunction() invocation into the
final field value.